feat(deps): Use drop-in pgzip instead of gzip (#178)
In my meager testing, this is about 10% faster.

If the [performance claims][1] are to believed, it will likely be a lot
faster in practice for Vela dependency caching, which will typically
handle more than 1 MB files.

I sought some improvement after noticing a 700 MB deps archive taking
30 seconds to restore/rebuild. This looks like a convenient, drop-in
replacement from the same author as the library already in use.

[1]: https://github.com/klauspost/pgzip#performance